+++ src/internet/model/ipv4-interface-address.h
@@ -101,16 +101,13 @@ class Ipv4InterfaceAddress
      * @param mask the network mask
      */
     void SetMask(Ipv4Mask mask);
+
     /**
      * @brief Get the network mask
      * @returns the network mask
      */
     Ipv4Mask GetMask() const;
-    /**
-     * @brief Set the broadcast address
-     * @param broadcast the broadcast address
-     */
-    void SetBroadcast(Ipv4Address broadcast);
+
     /**
      * @brief Get the broadcast address
      * @returns the broadcast address
@@ -157,8 +154,7 @@ class Ipv4InterfaceAddress
     Ipv4Address m_local; //!< Interface address
     // Note:  m_peer may be added in future when necessary
     // Ipv4Address m_peer;   // Peer destination address (in Linux:  m_address)
-    Ipv4Mask m_mask;         //!< Network mask
-    Ipv4Address m_broadcast; //!< Broadcast address
+    Ipv4Mask m_mask; //!< Network mask
 
     InterfaceAddressScope_e m_scope; //!< Address scope
     bool m_secondary;                //!< For use in multihoming
@@ -194,15 +190,15 @@ std::ostream& operator<<(std::ostream& os, const Ipv4InterfaceAddress& addr);
 inline bool
 operator==(const Ipv4InterfaceAddress& a, const Ipv4InterfaceAddress& b)
 {
-    return (a.m_local == b.m_local && a.m_mask == b.m_mask && a.m_broadcast == b.m_broadcast &&
-            a.m_scope == b.m_scope && a.m_secondary == b.m_secondary);
+    return (a.m_local == b.m_local && a.m_mask == b.m_mask && a.m_scope == b.m_scope &&
+            a.m_secondary == b.m_secondary);
 }
 
 inline bool
 operator!=(const Ipv4InterfaceAddress& a, const Ipv4InterfaceAddress& b)
 {
-    return (a.m_local != b.m_local || a.m_mask != b.m_mask || a.m_broadcast != b.m_broadcast ||
-            a.m_scope != b.m_scope || a.m_secondary != b.m_secondary);
+    return (a.m_local != b.m_local || a.m_mask != b.m_mask || a.m_scope != b.m_scope ||
+            a.m_secondary != b.m_secondary);
 }
 
 } // namespace ns3